Refactoring after removing Addressee and CtscContactDetails
